Changes:
 - adapted lpsconfcheck to use tool classes (together with Jeroen Keiren)

    This is an initial version, a prover tool class must still be added.

 - rewriter: removed deprecated functions: get_rewriter()
 - rewriter: added deprecated function: get_strategy()

    This function is added for use with the prover components that expect a
    strategy instead of a rewriter object.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6334 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/squadt/processor.cpp
@@ -846,9 +846,9 @@ namespace squadt {
      }
   
      std::streamsize xsputn(const char * s, std::streamsize n) {
-       m_crc_computer.process_bytes(s, n);
+       m_crc_computer.process_bytes(s, static_cast< size_t >(n));
   
-       pbump(n);
+       pbump(static_cast< int >(n));
   
        return n;
      }
